Finishing cleaning up the memory reduction commit...  Main fix was to add "delete d_archive" back to SimulationController.cc.
M    CCA/Components/SimulationController/SimulationController.cc

- Add back in "delete d_archive"... This still needs to be investigated.

M    CCA/Components/LoadBalancers/CostProfiler.cc
M    CCA/Components/LoadBalancers/DynamicLoadBalancer.cc

- Remove warning: numLevels is unsigned now.
- Cosmetics: coding standard, white space, {}s, etc

M    Core/Grid/Patch.cc
M    CCA/Components/Schedulers/DWDatabase.h
M    CCA/Components/Schedulers/MPIScheduler.cc
M    CCA/Components/Schedulers/OnDemandDataWarehouse.cc

- Cosmetics: line up variables, untabify, coding standard, white space, {}s, etc

M    CCA/Components/Schedulers/SchedulerCommon.cc

- Remove warning: numLevels is unsigned now.

+++ src/CCA/Components/Schedulers/SchedulerCommon.cc
@@ -1535,7 +1535,7 @@ SchedulerCommon::copyDataToNewGrid( const ProcessorGroup*,
     vector<ParticleSubset*> oldsubsets(d_sharedState->getNumMatls()), newsubsets(d_sharedState->getNumMatls());
 
     // If there is a level that didn't exist, we don't need to copy it
-    if ( newLevel->getIndex() >= oldDataWarehouse->getGrid()->numLevels() ) {
+    if ( newLevel->getIndex() >= (int)oldDataWarehouse->getGrid()->numLevels() ) {
       continue;
     }
